apps/backend/pyproject.toml
[build-system]
requires = ["setuptools>=69.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "comprasgov-backend"
version = "0.1.0"
description = "NEXORA unified backend with GraphQL gateway"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110,<0.112",
    "uvicorn[standard]>=0.29,<0.30",
    "pydantic-settings>=2.3,<3.0",
    "strawberry-graphql[fastapi]>=0.211,<0.213",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<8.0",
    "httpx>=0.26,<0.27",
]

[tool.black]
line-length = 99

[tool.isort]
profile = "black"
